refactor: getCars()를 InputView로 이동, 메서드 분리
