+++ src/main/java/racingcar/OutputView.java
@@ -1,0 +1,6 @@
+package racingcar;
+
+public class OutputView {
+
+
+}
